重构详细内容activity,添加分享功能，夜间模式功能

+++ app/src/main/java/com/example/windy/wind/database/cache/CacheManager.java
@@ -53,7 +53,9 @@ public class CacheManager implements ClearCacheInterface{
 
                             //删除详细内容缓存
                             ZhihuDailyContent content = mDb.zhihuDailyContentDao().query(item.getId());
-                            mDb.zhihuDailyContentDao().delete(content);
+                            if (content != null) {
+                                mDb.zhihuDailyContentDao().delete(content);
+                            }
                         }
                         hasCleared = true;
                     }
